Only delete target directories if overwrite flag is not set
This makes the behaviour for output directories match the behaviour for output files. 

In practice it is going to be a bad idea to specify the overwrite flag for anything outputting a directory, and Bee's API will encourage people away from this accordingly, but there are a few cases (like builds.zip unpacking) where we want it.