Fix file enumeration callback params consistency
On Win32, file enumeration was passing the entire 'scan path' - the path to the file from the root of the enumeration - to the callback, while on macOS/Linux, we were only passing the file name.

While we were only using this for signature calculation, it didn't make much difference - potentially we would have miscalculated signatures on POSIX if you moved a file between directories, we might not have noticed you moved it - but now that we are doing DynamicOuptutDirectories we need the paths to actually be correct and consistent across platforms, so that we do not end up thinking that the outputs were in the root directory.

Now that the enumeration code is always passing the 'scan path' to the callback, we no longer need to construct it ourselves in the DynamicOutputDirectories code, so revert the previous commit.