Replace frontend running with exit behaviour
If the frontend is out of date and needs to be regenerated, then instead of taking responsibility for running the frontend ourselves, just exit with the 'require frontend rerun' exit code. It is now the responsibility of the caller to recognise that exit code and regenerate the JSON DAG before running again.