Swallow action output that does not contain non-newline characters
Normally, we would swallow output if allowedOuputSubstrings were
specified for an action (and one of the patterns matched). But there
was a corner case for output that only contains newlines, then we would
not check the patterns and output was not being swallowed. Let's just go
ahead and always swallow this corner-case output.