exit early if there's an unreadable state file
If there is an unreadable state file, then tundra won't be able to
perform the stale output removal step, and we might produce incorrect
build results.

This can happen when the tundra binary is updated to a version with a
different magic number (tundra can only read frozen data files with the
same magic number).

Instead of quietly producing incorrect results, we should exit early
with an error message advising the user to clean the artifacts directory
and then to rebuild.