Replace WriteTextFile flag with ActionType enum
Using an enum instead of a flag allows us to add more actions more easily (such as file copy); it's also better than just adding more flags because a node should only be able to have one action type at a time, while flags leaves open the possibility of multiple flags being set at once.

Rather than introducing more storage for the action type, I decided to store it in the lower 8 bits of the 32 currently reserved for flags (of which we're only using 6, now). So, I also renamed 'm_Flags' to 'm_FlagsAndActionType' to indicate that two things are being stored there now. The values of the flags have been raised so that there is no need to shift when accessing them.