Add new 'ExecuteDirect' action
Usually we wrap all action invocations in a shell process. This allows us to have actions that have full shell shenanigans in play - redirection, piping, locating executables on the PATH, etc - but comes with a performance hit. A lot of the actions we take don't need this - we know the path to the executable and we are just going to launch it with some arguments - so introduce this new ActionType of `ExecuteDirect` that we can use for actions where we know we don't need to spend time kicking up a shell.